requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astor==0.8.1
attrs==23.1.0
black==21.4b2
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cycler==0.11.0
decorator==4.4.2
detectron2==0.6+cu113
fonttools==4.38.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0rc1
google-auth-oauthlib==0.4.6
grpcio==1.56.0
gym==0.10.9
habitat==0.1.5
habitat-sim==0.1.5
hydra-core==1.3.2
idna==3.4
ifcfg==0.24
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
llvmlite==0.39.1
magnum==0.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
moviepy==2.0.0.dev2
mypy-extensions==1.0.0
numba==0.56.4
numpy==1.21.6
numpy-quaternion==2022.4.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pip==22.3.1
platformdirs==3.8.0
portalocker==2.7.0
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydot==1.4.2
pyglet==2.0.8
pyparsing==3.1.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
tb-nightly==2.10.0a20220720
tensorboard==1.13.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.0+cu113
tqdm==4.65.0
typed-ast==1.5.4
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.2.3
wheel==0.38.4
yacs==0.1.8
zipp==3.15.0
